refactor(mcp): split enterprise.rs and cloud.rs into domain submodules
Break two large single-file tool modules into cohesive submodules:

enterprise.rs (3287 lines) -> enterprise/{mod,cluster,databases,rbac,observability}.rs
cloud.rs (1582 lines) -> cloud/{mod,subscriptions,account}.rs

Public API preserved via wildcard re-exports so all existing paths
(e.g. tools::enterprise::get_cluster) continue to resolve.
